feat: separate HashConsensus events with fixes
