+++ csm-alerts/src/services/EventsWatcher/events/index.ts
@@ -1,4 +1,5 @@
 export * from './assetRecoverer'
+export * from './hashconsensus'
 export * from './distributor'
 export * from './accounting'
 export * from './pausable'
